conf_packet_size is used as CONF_TYPE_INT, which is a 32bit type (fixes possible memory corruption).
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34367 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/muxer_mpeg.c
@@ -89,7 +89,7 @@ static const char *aspect_ratios[] = {
 };
 
 static char *conf_mux = "mpeg2";
-static uint16_t conf_packet_size = 0;		//dvd
+static uint32_t conf_packet_size = 0;		//dvd
 static uint32_t conf_muxrate = 0;		//kb/s
 static float conf_vaspect = 0;
 static float conf_vframerate = 0;
